+++ changes.md
@@ -1,3 +1,7 @@
+###  220720
+
+* Fehler im Text abspeichern gefixt.
+
 ###  220518
 
 * WordPress 6